+++ database/migrations/2024_02_29_192317_create_trix_rich_texts_table.php
@@ -13,24 +13,26 @@ class CreateTrixRichTextsTable extends Migration
      */
     public function up()
     {
-        Schema::create('trix_rich_texts', function (Blueprint $table) {
-            $table->unsignedBigInteger('id')->autoIncrement();
-            $table->string('field');
-            $table->morphs('model');
-            $table->text('content')->nullable();
-            $table->timestamps();
-        });
+        if (DB::connection()->getDriverName() == 'mysql')
+            Schema::create('trix_rich_texts', function (Blueprint $table) {
+                $table->unsignedBigInteger('id')->autoIncrement();
+                $table->string('field');
+                $table->morphs('model');
+                $table->text('content')->nullable();
+                $table->timestamps();
+            });
 
-        Schema::create('trix_attachments', function (Blueprint $table) {
-            $table->unsignedBigInteger('id')->autoIncrement();
-            $table->string('field');
-            $table->unsignedInteger('attachable_id')->nullable();
-            $table->string('attachable_type');
-            $table->string('attachment');
-            $table->string('disk');
-            $table->boolean('is_pending')->default(1);
-            $table->timestamps();
-        });
+        if (DB::connection()->getDriverName() == 'mysql')
+            Schema::create('trix_attachments', function (Blueprint $table) {
+                $table->unsignedBigInteger('id')->autoIncrement();
+                $table->string('field');
+                $table->unsignedInteger('attachable_id')->nullable();
+                $table->string('attachable_type');
+                $table->string('attachment');
+                $table->string('disk');
+                $table->boolean('is_pending')->default(1);
+                $table->timestamps();
+            });
     }
 
     /**
@@ -40,7 +42,9 @@ class CreateTrixRichTextsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('trix_attachments');
-        Schema::drop('trix_rich_texts');
+        if (DB::connection()->getDriverName() == 'mysql') {
+            Schema::drop('trix_attachments');
+            Schema::drop('trix_rich_texts');
+        }
     }
 }
